+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/CoreInstallersBundle.java
@@ -1,0 +1,39 @@
+package ru.vyarus.dropwizard.guice.module.installer;
+
+import ru.vyarus.dropwizard.guice.module.installer.feature.LifeCycleInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.feature.ManagedInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.feature.TaskInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.feature.admin.AdminFilterInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.feature.admin.AdminServletInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.feature.eager.EagerSingletonInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.feature.health.HealthCheckInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.feature.jersey.ResourceInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.feature.jersey.provider.JerseyProviderInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.feature.plugin.PluginInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.bundle.GuiceyBootstrap;
+import ru.vyarus.dropwizard.guice.module.installer.bundle.GuiceyBundle;
+
+/**
+ * Core installers bundle.
+ *
+ * @author Vyacheslav Rusakov
+ * @since 02.08.2015
+ */
+public class CoreInstallersBundle implements GuiceyBundle {
+
+    @Override
+    public void initialize(final GuiceyBootstrap bootstrap) {
+        bootstrap.installers(
+                LifeCycleInstaller.class,
+                ManagedInstaller.class,
+                JerseyProviderInstaller.class,
+                ResourceInstaller.class,
+                EagerSingletonInstaller.class,
+                HealthCheckInstaller.class,
+                TaskInstaller.class,
+                PluginInstaller.class,
+                AdminFilterInstaller.class,
+                AdminServletInstaller.class
+        );
+    }
+}