+++ src/test/groovy/ru/vyarus/dropwizard/guice/bundles/bootstrap/BootstrapMethodsTest.groovy
@@ -1,0 +1,50 @@
+package ru.vyarus.dropwizard.guice.bundles.bootstrap
+
+import com.google.inject.Inject
+import com.google.inject.Injector
+import com.google.inject.Key
+import io.dropwizard.Configuration
+import io.dropwizard.setup.Environment
+import ru.vyarus.dropwizard.guice.AbstractTest
+import ru.vyarus.dropwizard.guice.module.installer.feature.ManagedInstaller
+import ru.vyarus.dropwizard.guice.module.installer.feature.TaskInstaller
+import ru.vyarus.dropwizard.guice.module.installer.feature.jersey.ResourceInstaller
+import ru.vyarus.dropwizard.guice.module.installer.internal.FeaturesHolder
+import ru.vyarus.dropwizard.guice.support.TestConfiguration
+import ru.vyarus.dropwizard.guice.support.feature.DummyManaged
+import ru.vyarus.dropwizard.guice.support.feature.DummyResource
+import ru.vyarus.dropwizard.guice.support.feature.DummyTask
+import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyApp
+
+/**
+ * @author Vyacheslav Rusakov 
+ * @since 03.08.2015
+ */
+@UseGuiceyApp(GBootstrapApplication)
+class BootstrapMethodsTest extends AbstractTest {
+
+    @Inject
+    FeaturesHolder holder
+    @Inject
+    Injector injector
+
+    def "Check bootstrap config"() {
+
+        when: "application started"
+
+        then: "all registered installers found"
+        holder.installers.size() == 8
+
+        then: "task found"
+        holder.getFeatures(TaskInstaller) == [DummyTask]
+        injector.getExistingBinding(Key.get(DummyTask))
+
+        then: "resource found"
+        holder.getFeatures(ResourceInstaller) == [DummyResource]
+        injector.getExistingBinding(Key.get(DummyResource))
+
+        then: "managed found"
+        holder.getFeatures(ManagedInstaller) == [DummyManaged]
+        injector.getExistingBinding(Key.get(DummyManaged))
+    }
+}