+++ src/test/groovy/ru/vyarus/dropwizard/guice/bundles/manual/ManualBundlesRegistrationTest.groovy
@@ -1,0 +1,25 @@
+package ru.vyarus.dropwizard.guice.bundles.manual
+
+import com.google.inject.Inject
+import ru.vyarus.dropwizard.guice.AbstractTest
+import ru.vyarus.dropwizard.guice.module.installer.internal.FeaturesHolder
+import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyApp
+
+/**
+ * @author Vyacheslav Rusakov 
+ * @since 02.08.2015
+ */
+@UseGuiceyApp(ManualBundlesApplication)
+class ManualBundlesRegistrationTest extends AbstractTest {
+
+    @Inject
+    FeaturesHolder holder
+
+    def "Check manual bundle processed"() {
+
+        when: "application started"
+        then: "installers registered"
+        holder.installers.size() == 10
+
+    }
+}